Separate properties.go into separate files properties.go just a dumping ground for helper functions, it's a bit of a mess, this is the first step of cleaning it up
