+++ pkg/api/v1beta3/dynakube/extensions_props.go
@@ -1,0 +1,14 @@
+package dynakube
+
+const (
+	ExtensionsExecutionControllerStatefulsetName = "dynatrace-extensions-controller"
+	ExtensionsCollectorStatefulsetName           = "dynatrace-extensions-collector"
+)
+
+func (dk *DynaKube) IsExtensionsEnabled() bool {
+	return dk.Spec.Extensions.Prometheus.Enabled
+}
+
+func (dk *DynaKube) PrometheusEnabled() bool {
+	return dk.Spec.Extensions.Prometheus.Enabled
+}
