+++ pkg/api/v1beta3/dynakube/logmodule_props.go
@@ -1,0 +1,5 @@
+package dynakube
+
+func (dk *DynaKube) NeedsLogModule() bool {
+	return dk.Spec.LogModule.Enabled
+}
