+++ pkg/api/v1beta3/dynakube/metadata_enrichment_props.go
@@ -1,0 +1,13 @@
+package dynakube
+
+import (
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+)
+
+func (dk *DynaKube) MetadataEnrichmentEnabled() bool {
+	return dk.Spec.MetadataEnrichment.Enabled
+}
+
+func (dk *DynaKube) MetadataEnrichmentNamespaceSelector() *metav1.LabelSelector {
+	return &dk.Spec.MetadataEnrichment.NamespaceSelector
+}
